requirement.txt
Pillow>=10.3.0
PyMuPDF>=1.23.26
beautifulsoup4>=4.12.3
EbookLib>=0.18
patool>=1.12
pikepdf>=8.1.0
py7zr>=0.21.0
pypinyin>=0.50.0
rarfile>=4.1
tqdm>=4.66.1